Test reproducibility of an R Markdown file

Test reproducibility of an R Markdown file

Usage

isReproducible(filename, resetOptions = TRUE, ...)

Arguments

filename

Character. An R Markdown file to check for reproducibility

resetOptions

Boolean. Should all package options be reset to defaults? TRUE by default. This avoids problems if multiple checks on multiple documents with varying options are made in a row

...

Optional arguments passed down to rmarkdown::render()